+++ packages/modal/src/Modal.spec.tsx
@@ -1,25 +1,36 @@
-import React from 'react';
-import { render, screen } from '@testing-library/react';
+import React, { useState } from 'react';
+import {
+  fireEvent,
+  render,
+  waitForElementToBeRemoved,
+} from '@testing-library/react';
 import Modal from './Modal';
 
 const modalContent = 'Modal Content';
 
-function renderModal(props = {}) {
-  const utils = render(
-    <Modal data-testid="modal-test-id" {...props}>
+const ModalWrapper = ({
+  open: initialOpen,
+  ...props
+}: Partial<React.ComponentProps<typeof Modal>>) => {
+  const [open, setOpen] = useState(initialOpen);
+
+  return (
+    <Modal data-testid="modal-test-id" {...props} open={open} setOpen={setOpen}>
       {modalContent}
-    </Modal>,
+    </Modal>
   );
-  return utils;
+};
+
+function renderModal(props: Partial<React.ComponentProps<typeof Modal>> = {}) {
+  return render(<ModalWrapper {...props}>{modalContent}</ModalWrapper>);
 }
 
 describe('packages/modal', () => {
   describe('when the "open" prop is true', () => {
     test('renders modal to the DOM', () => {
-      const { getByTestId } = renderModal({ open: true });
-      const modal = getByTestId('modal-test-id');
+      const { getByRole } = renderModal({ open: true });
+      const modal = getByRole('dialog');
 
-      expect(modal).toBeInTheDocument();
       expect(modal).toBeVisible();
     });
 
@@ -29,54 +40,75 @@ describe('packages/modal', () => {
       expect(content).toBeVisible();
     });
 
-    test('closes modal when escape key is pressed', () => {
-      const { container } = renderModal({ open: true });
-      window.dispatchEvent(
-        new KeyboardEvent('keydown', { bubbles: true, key: 'Escape' }),
-      );
-      expect(container.innerHTML).toBe('');
+    test('closes modal when escape key is pressed', async () => {
+      const { getByRole } = renderModal({ open: true });
+
+      const modal = getByRole('dialog');
+      fireEvent.keyDown(document, { key: 'Escape', keyCode: 27 });
+
+      await waitForElementToBeRemoved(modal);
     });
 
-    test('when "shouldClose" prop returns false', () => {
-      const { getByTestId } = renderModal({
+    test('when "shouldClose" prop returns false', async () => {
+      const { getByRole } = renderModal({
         open: true,
         shouldClose: () => false,
       });
-      const modal = getByTestId('modal-test-id');
-      window.dispatchEvent(new MouseEvent('click', { bubbles: true }));
 
+      const modal = getByRole('dialog');
+      fireEvent.click(modal.parentElement!);
+
+      await expect(waitForElementToBeRemoved(modal)).rejects.toEqual(
+        Error('Timed out in waitForElementToBeRemoved.'),
+      );
       expect(modal).toBeVisible();
     });
 
-    test('when "shouldClose" returns true', () => {
-      const { container } = renderModal({
+    test('when "shouldClose" returns true', async () => {
+      const { getByRole } = renderModal({
         open: true,
         shouldClose: () => true,
       });
 
-      window.dispatchEvent(new MouseEvent('click', { bubbles: true }));
-      expect(container.innerHTML).toBe('');
+      const modal = getByRole('dialog');
+      fireEvent.click(modal.parentElement!);
+
+      await waitForElementToBeRemoved(modal);
     });
 
-    test('when "closeOnBackdropClick" is false', () => {
-      renderModal({ closeOnBackdropClick: false, open: true });
-      window.dispatchEvent(new MouseEvent('click', { bubbles: true }));
-      expect(screen.getByTestId('modal-test-id')).toBeInTheDocument();
+    test('when "closeOnBackdropClick" is false', async () => {
+      const { getByRole } = renderModal({
+        closeOnBackdropClick: false,
+        open: true,
+      });
+
+      const modal = getByRole('dialog');
+      fireEvent.click(modal.parentElement!);
+
+      await expect(waitForElementToBeRemoved(modal)).rejects.toEqual(
+        Error('Timed out in waitForElementToBeRemoved.'),
+      );
+      expect(modal).toBeVisible();
     });
-    test('when the "closeOnBackdropClick" is true', () => {
-      const { container } = renderModal({
+
+    test('when "closeOnBackdropClick" is true', async () => {
+      const { getByRole } = renderModal({
         closeOnBackdropClick: true,
         open: true,
       });
-      window.dispatchEvent(new MouseEvent('click', { bubbles: true }));
-      expect(container.innerHTML).toBe('');
+
+      const modal = getByRole('dialog');
+      fireEvent.click(modal.parentElement!);
+
+      await waitForElementToBeRemoved(modal);
     });
   });
 
   describe('when "open" prop is false', () => {
     test('does not render to the DOM', () => {
-      const { container } = renderModal({ open: false });
-      expect(container.innerHTML).toBe('');
+      const { queryByRole } = renderModal({ open: false });
+      const modal = queryByRole('dialog');
+      expect(modal).not.toBeInTheDocument();
     });
   });
 });