+++ packages/logo/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @leafygreen-ui/logo
 
+## 2.1.2
+
+### Patch Changes
+
+- 691eb05: Better support for UMD
+- Updated dependencies [691eb05]
+  - @leafygreen-ui/emotion@2.0.2
+  - @leafygreen-ui/lib@5.1.1
+  - @leafygreen-ui/palette@2.0.2
+
 ## 2.1.1
 
 ### Patch Changes