+++ packages/theme/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leafygreen-ui/theme
 
+## 2.0.1
+
+### Patch Changes
+
+- 691eb05: Better support for UMD
+
 ## 2.0.0
 
 ### Major Changes